+++ cmd/psweb/ln/lnd.go
@@ -3,14 +3,21 @@
 package ln
 
 import (
+	"bytes"
 	"context"
+	"encoding/base64"
+	"encoding/hex"
 	"errors"
 	"fmt"
 	"log"
 	"os"
 	"strconv"
+	"strings"
 
+	"peerswap-web/cmd/psweb/bitcoin"
 	"peerswap-web/cmd/psweb/config"
+
+	"github.com/btcsuite/btcd/wire"
 
 	"github.com/lightningnetwork/lnd/lnrpc"
 	"github.com/lightningnetwork/lnd/lnrpc/walletrpc"
@@ -19,6 +26,13 @@ import (
 	"google.golang.org/grpc/credentials"
 	"gopkg.in/macaroon.v2"
 )
+
+var internalLockId = []byte{
+	0xed, 0xe1, 0x9a, 0x92, 0xed, 0x32, 0x1a, 0x47,
+	0x05, 0xf8, 0xa1, 0xcc, 0xcc, 0x1d, 0x4f, 0x61,
+	0x82, 0x54, 0x5d, 0x4b, 0xb4, 0xfa, 0xe0, 0x8b,
+	0xd5, 0x93, 0x78, 0x31, 0xb7, 0xe3, 0x8f, 0x98,
+}
 
 func lndConnection() (*grpc.ClientConn, error) {
 	tlsCertPath := config.GetPeerswapLNDSetting("lnd.tlscertpath")
@@ -75,23 +89,6 @@ func GetClient() (lnrpc.LightningClient, func(), error) {
 	return lnrpc.NewLightningClient(conn), cleanup, nil
 }
 
-func SendCoins(client lnrpc.LightningClient, addr string, amount int64, feeRate uint64, sweepall bool, label string) (string, error) {
-	ctx := context.Background()
-	resp, err := client.SendCoins(ctx, &lnrpc.SendCoinsRequest{
-		Addr:        addr,
-		Amount:      amount,
-		SatPerVbyte: feeRate,
-		SendAll:     sweepall,
-		Label:       label,
-	})
-	if err != nil {
-		log.Println("SendCoins:", err)
-		return "", err
-	}
-
-	return resp.Txid, nil
-}
-
 func ConfirmedWalletBalance(client lnrpc.LightningClient) int64 {
 	ctx := context.Background()
 	resp, err := client.WalletBalance(ctx, &lnrpc.WalletBalanceRequest{})
@@ -125,6 +122,9 @@ func ListUnspent(_ lnrpc.LightningClient, list *[]UTXO, minConfs int32) error {
 			Address:       i.Address,
 			AmountSat:     i.AmountSat,
 			Confirmations: i.Confirmations,
+			TxidBytes:     i.Outpoint.GetTxidBytes(),
+			TxidStr:       i.Outpoint.GetTxidStr(),
+			OutputIndex:   i.Outpoint.GetOutputIndex(),
 		})
 	}
 
@@ -178,55 +178,6 @@ func GetAlias(nodeKey string) string {
 	return nodeInfo.Node.Alias
 }
 
-func BumpPeginFee(newFeeRate uint64) error {
-	client, cleanup, err := GetClient()
-	if err != nil {
-		return err
-	}
-	defer cleanup()
-
-	tx, err := getTransaction(client, config.Config.PeginTxId)
-	if err != nil {
-		return err
-	}
-
-	if len(tx.OutputDetails) == 1 {
-		return errors.New("peg-in transaction has no change output, not possible to CPFP")
-	}
-
-	outputIndex := uint32(999) // will fail if output not found
-	for _, output := range tx.OutputDetails {
-		if output.Amount != config.Config.PeginAmount {
-			outputIndex = uint32(output.OutputIndex)
-			break
-		}
-	}
-
-	ctx := context.Background()
-	conn, err := lndConnection()
-	if err != nil {
-		return err
-	}
-	cl := walletrpc.NewWalletKitClient(conn)
-
-	_, err = cl.BumpFee(ctx, &walletrpc.BumpFeeRequest{
-		Outpoint: &lnrpc.OutPoint{
-			TxidStr:     config.Config.PeginTxId,
-			OutputIndex: outputIndex,
-		},
-		SatPerVbyte: newFeeRate,
-	})
-
-	if err != nil {
-		log.Println("BumpFee:", err)
-		return err
-	}
-
-	log.Println("Fee bump successful")
-	conn.Close()
-	return nil
-}
-
 func GetRawTransaction(client lnrpc.LightningClient, txid string) (string, error) {
 	tx, err := getTransaction(client, txid)
 
@@ -237,44 +188,220 @@ func GetRawTransaction(client lnrpc.LightningClient, txid string) (string, error
 	return tx.RawTxHex, nil
 }
 
-func FindChildTx(client lnrpc.LightningClient) string {
-	tx, err := getTransaction(client, config.Config.PeginTxId)
-	if err != nil {
-		return ""
-	}
-
-	if len(tx.OutputDetails) == 1 {
-		return ""
-	}
-
-	outputIndex := uint32(999) // will fail if output not found
-	for _, output := range tx.OutputDetails {
-		if output.Amount != config.Config.PeginAmount {
-			outputIndex = uint32(output.OutputIndex)
-			break
-		}
-	}
-
+// utxos: ["txid:index", ....]
+// return rawTx hex string, final output amount, error
+func PrepareRawTxWithUtxos(utxos *[]string, addr string, amount int64, feeRate uint64, subtractFeeFromAmount bool) (string, int64, error) {
 	ctx := context.Background()
-
-	// find the child tx in the wallet
-	resp, err := client.GetTransactions(ctx, &lnrpc.GetTransactionsRequest{})
+	conn, err := lndConnection()
 	if err != nil {
-		return ""
+		return "", 0, err
+	}
+	cl := walletrpc.NewWalletKitClient(conn)
+
+	outputs := make(map[string]uint64)
+	finalAmount := amount
+
+	if subtractFeeFromAmount {
+		if len(*utxos) == 0 {
+			return "", 0, errors.New("at least one unspent output must be selected to send funds without change")
+		}
+		// give no output address,
+		// let fundpsbt change back to the wallet to find the fee amount
+	} else {
+		outputs[addr] = uint64(amount)
 	}
 
-	txid := ""
-	timestamp := int64(0)
-	vin := config.Config.PeginTxId + ":" + strconv.FormatUint(uint64(outputIndex), 10)
-	for _, tx := range resp.Transactions {
-		for _, in := range tx.PreviousOutpoints {
-			// find the latest child spending our output
-			if in.Outpoint == vin && tx.TimeStamp > timestamp {
-				txid = tx.TxHash
-				timestamp = tx.TimeStamp
+	psbtBytes, err := fundPsbt(cl, utxos, outputs, feeRate)
+	if err != nil {
+		log.Println("FundPsbt:", err)
+		return "", 0, err
+	}
+
+	if subtractFeeFromAmount {
+		// replace output with correct address and amount
+		psbtString := base64.StdEncoding.EncodeToString(psbtBytes)
+		fee, err := bitcoin.GetFeeFromPsbt(psbtString)
+		if err != nil {
+			return "", 0, err
+		}
+		// reduce output amount by fee and small haircut to go around the LND bug
+		for haircut := int64(0); haircut <= 100; haircut += 5 {
+			err := releaseOutputs(cl, utxos)
+			if err != nil {
+				return "", 0, err
+			}
+
+			finalAmount = int64(amount) - toSats(fee) - haircut
+			if finalAmount < 0 {
+				finalAmount = 0
+			}
+			outputs[addr] = uint64(finalAmount)
+			psbtBytes, err = fundPsbt(cl, utxos, outputs, feeRate)
+			if err == nil {
+				break
 			}
 		}
+		if err != nil {
+			log.Println("FundPsbt:", err)
+			return "", 0, err
+		}
 	}
 
-	return txid
+	// sign psbt
+	res2, err := cl.FinalizePsbt(ctx, &walletrpc.FinalizePsbtRequest{
+		FundedPsbt: psbtBytes,
+	})
+
+	if err != nil {
+		log.Println("FinalizePsbt:", err)
+		releaseOutputs(cl, utxos)
+		return "", 0, err
+	}
+
+	rawTx := res2.GetRawFinalTx()
+
+	return hex.EncodeToString(rawTx), finalAmount, nil
+}
+
+func releaseOutputs(cl walletrpc.WalletKitClient, utxos *[]string) error {
+	ctx := context.Background()
+	for _, i := range *utxos {
+		parts := strings.Split(i, ":")
+
+		index, err := strconv.Atoi(parts[1])
+		if err != nil {
+			log.Println("releaseOutputs:", err)
+			break
+		}
+
+		_, err = cl.ReleaseOutput(ctx, &walletrpc.ReleaseOutputRequest{
+			Id: internalLockId,
+			Outpoint: &lnrpc.OutPoint{
+				TxidStr:     parts[0],
+				OutputIndex: uint32(index),
+			},
+		})
+
+		if err != nil {
+			log.Println("ReleaseOutput:", err)
+			return err
+		}
+	}
+	return nil
+}
+
+func fundPsbt(cl walletrpc.WalletKitClient, utxos *[]string, outputs map[string]uint64, feeRate uint64) ([]byte, error) {
+	ctx := context.Background()
+	var inputs []*lnrpc.OutPoint
+
+	for _, i := range *utxos {
+		parts := strings.Split(i, ":")
+
+		index, err := strconv.Atoi(parts[1])
+		if err != nil {
+			return nil, err
+		}
+
+		inputs = append(inputs, &lnrpc.OutPoint{
+			TxidStr:     parts[0],
+			OutputIndex: uint32(index),
+		})
+	}
+
+	res, err := cl.FundPsbt(ctx, &walletrpc.FundPsbtRequest{
+		Template: &walletrpc.FundPsbtRequest_Raw{
+			Raw: &walletrpc.TxTemplate{
+				Inputs:  inputs,
+				Outputs: outputs,
+			},
+		},
+		Fees: &walletrpc.FundPsbtRequest_SatPerVbyte{
+			SatPerVbyte: feeRate,
+		},
+		MinConfs:         int32(1),
+		SpendUnconfirmed: false,
+		ChangeType:       1, //P2TR
+	})
+
+	if err != nil {
+		return nil, err
+	}
+
+	return res.GetFundedPsbt(), nil
+}
+
+func RbfPegin(feeRate uint64) (string, int64, error) {
+	client, cleanup, err := GetClient()
+	if err != nil {
+		return "", 0, err
+	}
+	defer cleanup()
+
+	rawTx, err := GetRawTransaction(client, config.Config.PeginTxId)
+	if err != nil {
+		return "", 0, err
+	}
+
+	decodedTx, err := bitcoin.DecodeRawTransaction(rawTx)
+	if err != nil {
+		return "", 0, err
+	}
+
+	var utxos []string
+
+	for _, input := range decodedTx.Vin {
+		vin := input.TXID + ":" + strconv.FormatUint(uint64(input.Vout), 10)
+		utxos = append(utxos, vin)
+	}
+
+	conn, err := lndConnection()
+	if err != nil {
+		return "", 0, err
+	}
+	cl := walletrpc.NewWalletKitClient(conn)
+
+	err = releaseOutputs(cl, &utxos)
+	if err != nil {
+		return "", 0, err
+	}
+
+	return PrepareRawTxWithUtxos(
+		&utxos,
+		config.Config.PeginAddress,
+		config.Config.PeginAmount,
+		feeRate,
+		len(decodedTx.Vout) == 1)
+}
+
+func PublishTransaction(rawTx string, label string) (string, error) {
+	conn, err := lndConnection()
+	if err != nil {
+		return "", err
+	}
+	cl := walletrpc.NewWalletKitClient(conn)
+	ctx := context.Background()
+
+	tx, err := hex.DecodeString(rawTx)
+	if err != nil {
+		return "", err
+	}
+
+	// Deserialize the transaction to get the transaction hash.
+	msgTx := &wire.MsgTx{}
+	txReader := bytes.NewReader(tx)
+	if err := msgTx.Deserialize(txReader); err != nil {
+		return "", err
+	}
+
+	req := &walletrpc.Transaction{
+		TxHex: tx,
+		Label: label,
+	}
+
+	_, err = cl.PublishTransaction(ctx, req)
+	if err != nil {
+		log.Println("PublishTransaction:", err)
+		return "", err
+	}
+	return msgTx.TxHash().String(), nil
 }